v3.6.0: fix Float→Double cast for owwVad RMS/ZCR
Kotlin does not implicitly widen Float to Double for method
arguments (only for arithmetic). The v3.6.0 commit called
'params.putDouble("rms", rms)' with rms: Float, which fails
to compile against the real react-native WritableMap API
(putDouble takes Double only). Discovered by locally
compiling v3.6.0 against a stub of WritableMap with
putDouble(k: String, v: Double).

All other owwVad-event-scoring call sites in this file
(wakeScore, exitScore, sendScore) already had explicit
.toDouble() — this commit brings the new owwVad calls into
line with the existing pattern.

+++ android/app/src/main/java/com/cyberclawmobile/WakeWordModule.kt
@@ -1270,8 +1270,8 @@ class WakeWordModule(private val reactContext: ReactApplicationContext) :
                                     lastVadAt = now
                                     handler.post {
                                         val params = Arguments.createMap()
-                                        params.putDouble("rms", rms)
-                                        params.putDouble("zcr", zcr)
+                                        params.putDouble("rms", rms.toDouble())
+                                        params.putDouble("zcr", zcr.toDouble())
                                         reactContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
                                             .emit("owwVad", params)
                                     }
